fix common refactor pagination make it resusable
